feat: smooth scroll and refactor compliment component
