+++ src/main.rs
@@ -34,10 +34,10 @@ async fn main() -> std::io::Result<()> {
                 )),
             }))
             .service(actix_files::Files::new("/_static", "./static").show_files_listing())
-            .service(routes::compliments)
             .service(routes::landing_page)
             .service(routes::articles_page)
             .service(routes::article_page)
+            .service(routes::compliments::compliments) 
     })
     .bind(("0.0.0.0", port.parse::<u16>().unwrap()))?
     .run()
