+++ templates/articles_page.html
@@ -1,76 +1,1 @@
-{% extends "base.html" %} {% block title %}gustavo lee.{% endblock %} {% block
-content %}
-
-<div class="flex flex-col w-full items-center">
-  <div
-    class="flex flex-col gap-16 w-2/3 justify-center items-center bg-white border-x border-neutral-800/10"
-  >
-    <!-- Navbar -->
-    <nav class="flex justify-center items-center p-4">
-      <div class="flex justify-center items-center text-gray-900">
-        <div
-          class="font-bold border rounded-l-2xl border-neutral-800/10 p-2 shadow"
-        >
-          <a href="/" class="hover:text-orange-600 mx-2">Yo</a>
-          <a href="/articles" class="hover:text-orange-600 mx-2">Artículos</a>
-          <a href="#" class="hover:text-orange-600 mx-2">Trayectoria</a>
-          <a href="#" class="hover:text-orange-600 mx-2">Contacto</a>
-        </div>
-      </div>
-    </nav>
-
-    <!-- Introducción con foto, titulo y resumen -->
-    <div class="flex flex-col md:flex-row items-center justify-center">
-      <div class="w-1/3">
-        <img
-          src="https://avatars.githubusercontent.com/u/33813160?v=4"
-          alt="Gustavo Lee"
-          class="rounded-full border-2 border-neutral-800 w-52 h-52 mx-auto"
-        />
-      </div>
-      <div class="w-2/3">
-        <h2 class="text-2xl font-bold text-gray-900">Ingeniero de software</h2>
-        <p class="text-gray-700">
-          Desarrollador de software con experiencia en el desarrollo de
-          aplicaciones web y móviles. Me apasiona la tecnología y la
-          programación. Siempre estoy en busca de nuevos retos y oportunidades
-          para seguir aprendiendo.
-        </p>
-      </div>
-    </div>
-    <!-- Carusel de proyectos -->
-    <div>
-      <h2 class="text-2xl font-bold text-gray-900">Trayectoria</h2>
-      <div class="flex flex-row gap-4 justify-center items-center">
-        <div class="w-1/3">
-          <img
-            src="https://via.placeholder.com/600"
-            alt="Trayectoria 1"
-            class="rounded-lg w-64 h-64 mx-auto"
-          />
-          <p class="text-gray-700">Trayectoria 1</p>
-        </div>
-        <div class="w-1/3">
-          <img
-            src="https://via.placeholder.com/600"
-            alt="Trayectoria 2"
-            class="rounded-lg w-64 h-64 mx-auto"
-          />
-          <p class="text-gray-700">Trayectoria 2</p>
-        </div>
-        <div class="w-1/3">
-          <img
-            src="https://via.placeholder.com/600"
-            alt="Trayectoria 3"
-            class="rounded-lg w-64 h-64 mx-auto"
-          />
-          <p class="text-gray-700">Trayectoria 3</p>
-        </div>
-      </div>
-    </div>
-
-    <!-- Link a redes, linkdin y github -->
-  </div>
-</div>
-
-{% endblock %}
+{% block content %} {% endblock %}
