+++ templates/compliment.html
@@ -1,0 +1,13 @@
+{% block content %}
+
+<span
+  id="compliment"
+  class="inline-block font-migra-regular text-3xl xl:text-4xl pl-1 leading-none fade fade-in"
+  hx-target="this"
+  hx-get="/compliments/{{next_compliment_id}}"
+  hx-swap="outerHTML transition:true"
+  hx-trigger="every 3s"
+  >{{compliment}}</span
+>
+
+{% endblock %}
